Moved Subject extensions to its own file
