+++ lib/expectations_runner.rb
@@ -2,25 +2,7 @@ require 'mumukit'
 require 'mumukit/inspection'
 require 'stones-spec'
 
-module StonesSpec::Subject
-  class Procedure
-    def ast_regexp
-      /AST\(procedure\s*#{@name}/
-    end
-  end
-
-  class Function
-    def ast_regexp
-      /AST\(function\s*#{@name}/
-    end
-  end
-
-  module Program
-    def self.ast_regexp
-      /AST\(entrypoint\s*program/
-    end
-  end
-end
+require_relative 'subject_extensions'
 
 module EvalExpectationsOnAST
   def eval_in_gobstones(binding, ast)
